requirements.txt
kafka-python==2.0.2
redis==5.0.1
requests==2.31.0
fastapi==0.115.6
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
flask==3.0.2
pytest==8.0.0
aiohttp==3.9.3
PyYAML==6.0.1 
pytest-mock==3.12.0
asyncpg==0.29.0
httpx>=0.24.0
pytest-asyncio
